+++ fake-cucumber/javascript/test/CucumberTest.ts
@@ -37,6 +37,7 @@ describe('Cucumber', () => {
       gherkinMessageList,
       supportCode.parameterTypes,
       supportCode.stepDefinitions,
+      supportCode.undefinedParameterTypes,
       supportCode.beforeHooks,
       supportCode.afterHooks,
       gherkinQuery,