+++ gherkin/c/src/utf8_source.h
@@ -17,7 +17,7 @@ typedef void (*utf8_source_delete_function) (Utf8Source*);
 
 struct Utf8Source {
     utf8_source_read_function read;
-    utf8_source_delete_function delete;
+    utf8_source_delete_function del;
 };
 
 unsigned char Utf8Source_read(Utf8Source* utf8_source);