+++ json-formatter/go/test_step.go
@@ -4,7 +4,7 @@ import (
 	"encoding/base64"
 	"errors"
 	"fmt"
-	"github.com/cucumber/messages-go/v9"
+	"github.com/cucumber/messages-go/v10"
 	"strings"
 )
 
@@ -15,9 +15,10 @@ type TestStep struct {
 	PickleStep      *messages.Pickle_PickleStep
 	Step            *messages.GherkinDocument_Feature_Step
 	StepDefinitions []*messages.StepDefinition
-	Result          *messages.TestResult
+	Result          *messages.TestStepResult
 	Background      *messages.GherkinDocument_Feature_Background
 	Attachments     []*messages.Attachment
+	ExampleRow      *messages.GherkinDocument_Feature_TableRow
 }
 
 func ProcessTestStepFinished(testStepFinished *messages.TestStepFinished, lookup *MessageLookup) (error, *TestStep) {
@@ -45,7 +46,7 @@ func ProcessTestStepFinished(testStepFinished *messages.TestStepFinished, lookup
 		return nil, &TestStep{
 			TestCaseID: testCase.Id,
 			Hook:       hook,
-			Result:     testStepFinished.TestResult,
+			Result:     testStepFinished.TestStepResult,
 		}
 	}
 
@@ -59,6 +60,11 @@ func ProcessTestStepFinished(testStepFinished *messages.TestStepFinished, lookup
 		return errors.New("No pickleStep for " + testStep.PickleStepId), nil
 	}
 
+	var exampleRow *messages.GherkinDocument_Feature_TableRow
+	if len(pickle.AstNodeIds) > 1 {
+		exampleRow = lookup.LookupExampleRow(pickle.AstNodeIds[1])
+	}
+
 	var background *messages.GherkinDocument_Feature_Background
 	scenarioStep := lookup.LookupStep(pickleStep.AstNodeIds[0])
 	if scenarioStep != nil {
@@ -70,7 +76,8 @@ func ProcessTestStepFinished(testStepFinished *messages.TestStepFinished, lookup
 		Step:            lookup.LookupStep(pickleStep.AstNodeIds[0]),
 		Pickle:          pickle,
 		PickleStep:      pickleStep,
-		Result:          testStepFinished.TestResult,
+		ExampleRow:      exampleRow,
+		Result:          testStepFinished.TestStepResult,
 		StepDefinitions: lookup.LookupStepDefinitions(testStep.StepDefinitionIds),
 		Background:      background,
 		Attachments:     lookup.LookupAttachments(testStepFinished.TestStepId),
@@ -98,6 +105,10 @@ func TestStepToJSON(step *TestStep) *jsonStep {
 	}
 
 	location := makeLocation(step.Pickle.Uri, step.Step.Location.Line)
+	if step.ExampleRow != nil {
+		location = makeLocation(step.Pickle.Uri, step.ExampleRow.Location.Line)
+	}
+
 	if len(step.StepDefinitions) == 1 {
 		location = makeLocation(
 			step.StepDefinitions[0].SourceReference.Uri,
@@ -118,6 +129,7 @@ func TestStepToJSON(step *TestStep) *jsonStep {
 			Duration:     duration,
 		},
 		Embeddings: makeEmbeddings(step.Attachments),
+		Output:     makeOutput(step.Attachments),
 	}
 
 	docString := step.Step.GetDocString()
@@ -148,8 +160,10 @@ func TestStepToJSON(step *TestStep) *jsonStep {
 }
 
 func makeEmbeddings(attachments []*messages.Attachment) []*jsonEmbedding {
-	jsonEmbeddings := make([]*jsonEmbedding, len(attachments))
-	for index, attachment := range attachments {
+	embeddableAttachments := filterAttachments(attachments, isEmbeddable)
+	jsonEmbeddings := make([]*jsonEmbedding, len(embeddableAttachments))
+
+	for index, attachment := range embeddableAttachments {
 		var data []byte
 		if attachment.GetBinary() != nil {
 			data = attachment.GetBinary()
@@ -163,6 +177,35 @@ func makeEmbeddings(attachments []*messages.Attachment) []*jsonEmbedding {
 	}
 
 	return jsonEmbeddings
+}
+
+func makeOutput(attachments []*messages.Attachment) []string {
+	outputAttachments := filterAttachments(attachments, isOutput)
+	output := make([]string, len(outputAttachments))
+
+	for index, attachment := range outputAttachments {
+		output[index] = attachment.GetText()
+	}
+
+	return output
+}
+
+func filterAttachments(attachments []*messages.Attachment, filter func(*messages.Attachment) bool) []*messages.Attachment {
+	matches := make([]*messages.Attachment, 0)
+	for _, attachment := range attachments {
+		if filter(attachment) {
+			matches = append(matches, attachment)
+		}
+	}
+	return matches
+}
+
+func isEmbeddable(attachment *messages.Attachment) bool {
+	return !isOutput(attachment)
+}
+
+func isOutput(attachment *messages.Attachment) bool {
+	return attachment.GetMediaType() == "text/x.cucumber.log+plain"
 }
 
 func makeLocation(file string, line uint32) string {