+++ json-formatter/CHANGELOG.md
@@ -11,11 +11,21 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 
 ### Added
 
+* Handle text added with `log`.
+
 ### Changed
 
 ### Deprecated
 
 ### Removed
+
+### Fixed
+
+## [3.0.0] - 2020-02-14
+
+### Changed
+
+Upgrade messages
 
 ### Fixed
 
@@ -52,7 +62,8 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 * Initial release
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/cucumber/compare/json-formatter/v2.2.0...master
+[Unreleased]: https://github.com/cucumber/cucumber/compare/json-formatter/v3.0.0...master
+[3.0.0]:      https://github.com/cucumber/cucumber/compare/json-formatter/v2.2.0...json-formatter/v3.0.0
 [2.2.0]:      https://github.com/cucumber/cucumber/compare/json-formatter/v2.1.0...json-formatter/v2.2.0
 [2.1.0]:      https://github.com/cucumber/cucumber/compare/json-formatter/v2.0.0...json-formatter/v2.1.0
 [2.0.0]:      https://github.com/cucumber/cucumber/compare/json-formatter/v1.0.0...json-formatter/v2.0.0