+++ messages/CHANGELOG.md
@@ -17,6 +17,39 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+* [JavaScript] make stream readable/writable object modes explicit. Better buffer handling. 
+* [Ruby] Don't depend on the `json` gem, since it only installs on Windows with
+  a C compiler toolchain. The `json` gem is part of the Ruby stdlib.
+* [Go] Increase max size of a JSON message to 10Mb
+  ([#901](https://github.com/cucumber/cucumber/issues/901)
+   [#903](https://github.com/cucumber/cucumber/pull/903)
+   [aslakhellesoy])
+
+## [10.0.1] - 2020-02-13
+
+### Fixed
+
+* The 10.0.0 release failed.
+
+## [10.0.0] - 2020-02-13
+
+### Added
+
+* Add `UndefinedParameterType`
+  ([#890](https://github.com/cucumber/cucumber/pull/890)
+   [aslakhellesoy])
+* Add `TestCaseFinished#message`
+  ([#890](https://github.com/cucumber/cucumber/pull/890)
+   [aslakhellesoy])
+
+### Changed
+
+* Rename `TestResult` to `TestStepResult`
+  ([#890](https://github.com/cucumber/cucumber/pull/890)
+   [aslakhellesoy])
+
+### Fixed
+
 * Ruby: Fixed rounding errors in `TimeConversion`
 
 ## [9.0.3] - 2020-01-10
@@ -244,7 +277,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Protobuf messages for Go, Java, JavaScript, TypeScript and Ruby
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/cucumber/compare/messages/v9.0.3...master
+[Unreleased]: https://github.com/cucumber/cucumber/compare/messages/v10.0.1...master
+[10.0.1]:      https://github.com/cucumber/cucumber/compare/cucumber-messages/v10.0.0...messages/v10.0.1
+[10.0.0]:      https://github.com/cucumber/cucumber/compare/cucumber-messages/v9.0.3...messages/v10.0.0
 [9.0.3]:      https://github.com/cucumber/cucumber/compare/cucumber-messages/v9.0.2...messages/v9.0.3
 [9.0.2]:      https://github.com/cucumber/cucumber/compare/cucumber-messages/v9.0.1...messages/v9.0.2
 [9.0.1]:      https://github.com/cucumber/cucumber/compare/cucumber-messages/v9.0.0...messages/v9.0.1