+++ react/CHANGELOG.md
@@ -9,8 +9,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-* Display attachments (`image/*` and `text/*` media types)
-
 ### Changed
 
 ### Deprecated
@@ -18,6 +16,28 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+
+## [4.1.1] - 2020-02-28
+
+### Fixed
+
+* Fix broken 4.1.0 release
+
+## [4.1.0] - 2020-02-28
+
+### Added
+
+* Add compiled stylesheets to package
+
+## [4.0.0] - 2020-02-14
+
+### Added
+
+* Display attachments (`image/*` and `text/*` media types)
+
+### Changed
+
+* Upgraded messages and query
 
 ## [3.3.0] - 2020-01-22
 
@@ -84,7 +104,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * First release
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/cucumber/compare/react/v3.3.0...master
+[Unreleased]: https://github.com/cucumber/cucumber/compare/react/v4.1.1...master
+[4.1.1]:      https://github.com/cucumber/cucumber/compare/cucumber-react/v4.1.0...react/v4.1.1
+[4.1.0]:      https://github.com/cucumber/cucumber/compare/cucumber-react/v4.0.0...react/v4.1.0
+[4.0.0]:      https://github.com/cucumber/cucumber/compare/cucumber-react/v3.3.0...react/v4.0.0
 [3.3.0]:      https://github.com/cucumber/cucumber/compare/cucumber-react/v3.2.0...react/v3.3.0
 [3.2.0]:      https://github.com/cucumber/cucumber/compare/cucumber-react/v3.1.0...react/v3.2.0
 [3.1.0]:      https://github.com/cucumber/cucumber/compare/cucumber-react/v3.0.0...cucumber-react/v3.1.0